requirements.txt
openai
python-dotenv
azure-ai-formrecognizer==3.2.0
pandas==2.1.3
ipython
openpyxl
faiss-cpu